refactor: extract smaller functions from bite api service
