+++ libs/bite-tribe/api/src/lib/bite-api/utils/constants.ts
@@ -1,0 +1,2 @@
+export const BITE_COLLECTION = 'bites';
+export const DEFAULT_SEARCH_RADIUS_IN_M = 15 * 1000;
