+++ libs/bite-tribe/api/src/lib/review-api.service.ts
@@ -8,8 +8,8 @@ import {
   switchMap,
 } from 'rxjs';
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
-import { BITE_COLLECTION } from './bite-api/bite-api.service';
 import { User } from '@capacitor-firebase/authentication/dist/esm/definitions';
+import { BITE_COLLECTION } from './bite-api/utils/constants';
 
 const REVIEW_COLLECTION = 'reviews';
